updated term api e2e cypress test
